+++ dashboard/migrations/0001_initial.py
@@ -31,7 +31,7 @@ class Migration(migrations.Migration):
             },
         ),
         migrations.CreateModel(
-            name='LoanData',
+            name='LW321',
             fields=[
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('periode', models.CharField(db_index=True, max_length=30)),